requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2

pip~=21.2.3
numpy~=1.21.4
certifi~=2021.10.8
urllib3~=1.26.7
six~=1.16.0
idna~=3.3
setuptools~=57.4.0
djangorestframework~=3.12.4
requests~=2.26.0